more algorithm verification
