+++ python/pygfdm/utils.py
@@ -59,3 +59,17 @@ def get_zero_f_data(k, K, M):
     # data = np.tile(data, M)
     data = np.repeat(data, M)
     return data
+
+
+# function adopted from scikit-commpy. Separated noise variance calculation and noise vector generation.
+def calculate_awgn_noise_variance(input_signal, snr_dB, rate=1.0):
+    avg_energy = np.sum(input_signal * input_signal) / len(input_signal)
+    snr_linear = 10. ** (snr_dB / 10.0)
+    noise_variance = avg_energy/(2*rate*snr_linear)
+    return noise_variance
+
+
+# corresponds to 'calculate_awgn_noise_variance.
+def get_complex_noise_vector(nsamples, noise_variance):
+    return (np.sqrt(noise_variance) * np.random.randn(nsamples)) + (np.sqrt(noise_variance) * np.random.randn(nsamples) * 1j)
+
